+++ history/contrib/loaddata/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class HistoryLoaddataConfig(AppConfig):
+    name = "history.contrib.loaddata"
+    label = "history_loaddata"