+++ history/contrib/migrate/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class HistoryMigrateConfig(AppConfig):
+    name = "history.contrib.migrate"
+    label = "history_migrate"